requirements.txt
spacy==3.7.4
sentence-transformers==2.7.0
transformers==4.40.0
torch==2.3.0
nltk==3.8.1
datasets==2.19.0
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2
pdfminer.six==20221105
python-docx==1.1.2
pypdf2==3.0.1
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
python-multipart==0.0.9
pyyaml==6.0.1
loguru==0.7.2
requests==2.31.0
tqdm==4.66.4
python-dotenv==1.0.1
pytest==8.2.0
httpx==0.27.0